+++ internal/install/execution/output_parser.go
@@ -32,3 +32,10 @@ func (op *OutputParser) Metadata() map[string]string {
 	}
 	return nil
 }
+
+func (op *OutputParser) FailedRecipeOutput() string {
+	if val, ok := op.output["FailedRecipeOutput"]; ok {
+		return val.(string)
+	}
+	return ""
+}